+++ CHANGELOG.md
@@ -10,6 +10,8 @@
         ([#3120](https://github.com/Automattic/pocket-casts-android/pull/3120))
     *   Improve connection when some podcasts failed to play or download
         ([#3180](https://github.com/Automattic/pocket-casts-android/pull/3180))
+    *   Add new podcast subscriptions to the top instead of bottom for Data Added podcast sort order 
+        ([#3192](https://github.com/Automattic/pocket-casts-android/pull/3192)) 
 *   Bug Fixes
     *   Use red color for the notification icons
         ([#3154](https://github.com/Automattic/pocket-casts-android/pull/3154))